+++ src/Sulu/Component/Content/SmartContent/ContentDataProvider.php
@@ -16,7 +16,6 @@ use PHPCR\SessionInterface;
 use ProxyManager\Factory\LazyLoadingValueHolderFactory;
 use ProxyManager\Proxy\LazyLoadingInterface;
 use Sulu\Bundle\ContentBundle\Document\PageDocument;
-use Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStore;
 use Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStoreInterface;
 use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\Content\Query\ContentQueryBuilderInterface;
@@ -25,6 +24,7 @@ use Sulu\Component\DocumentManager\DocumentManagerInterface;
 use Sulu\Component\SmartContent\ArrayAccessItem;
 use Sulu\Component\SmartContent\Configuration\Builder;
 use Sulu\Component\SmartContent\Configuration\ProviderConfigurationInterface;
+use Sulu\Component\SmartContent\DataProviderAliasInterface;
 use Sulu\Component\SmartContent\DataProviderInterface;
 use Sulu\Component\SmartContent\DataProviderResult;
 use Sulu\Component\SmartContent\DatasourceItem;
@@ -32,7 +32,7 @@ use Sulu\Component\SmartContent\DatasourceItem;
 /**
  * DataProvider for content.
  */
-class ContentDataProvider implements DataProviderInterface
+class ContentDataProvider implements DataProviderInterface, DataProviderAliasInterface
 {
     /**
      * @var ContentQueryBuilderInterface
@@ -65,7 +65,7 @@ class ContentDataProvider implements DataProviderInterface
     private $session;
 
     /**
-     * @var ReferenceStore
+     * @var ReferenceStoreInterface
      */
     private $referenceStore;
 
@@ -206,16 +206,7 @@ class ContentDataProvider implements DataProviderInterface
 
         $items = $this->decorateDataItems($items, $options['locale']);
 
-        return new DataProviderResult(
-            $items,
-            $hasNextPage,
-            array_map(
-                function (ContentDataItem $item) {
-                    return $item->getId();
-                },
-                $items
-            )
-        );
+        return new DataProviderResult($items, $hasNextPage);
     }
 
     /**
@@ -239,16 +230,7 @@ class ContentDataProvider implements DataProviderInterface
         );
         $items = $this->decorateResourceItems($items, $options['locale']);
 
-        return new DataProviderResult(
-            $items,
-            $hasNextPage,
-            array_map(
-                function (ArrayAccessItem $item) {
-                    return $item->getId();
-                },
-                $items
-            )
-        );
+        return new DataProviderResult($items, $hasNextPage);
     }
 
     /**
@@ -280,11 +262,18 @@ class ContentDataProvider implements DataProviderInterface
         $properties = array_key_exists('properties', $propertyParameter) ?
             $propertyParameter['properties']->getValue() : [];
 
+        $excluded = $filters['excluded'];
+        if (array_key_exists('exclude_duplicates', $propertyParameter)
+            && $propertyParameter['exclude_duplicates']->getValue()
+        ) {
+            $excluded = array_merge($excluded, $this->referenceStore->getAll());
+        }
+
         $this->contentQueryBuilder->init(
             [
                 'config' => $filters,
                 'properties' => $properties,
-                'excluded' => $filters['excluded'],
+                'excluded' => $excluded,
                 'published' => !$this->showDrafts,
             ]
         );
@@ -418,5 +407,13 @@ class ContentDataProvider implements DataProviderInterface
                 return true;
             }
         );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getAlias()
+    {
+        return 'content';
     }
 }